Moved all of the add and remove URL logic to the background script
When I started implementing the blocked URL list with adding and removing functionality, I added all of this logic to the popup script. However, the problem with this is that the background script was never notified of any changes and could not update the list and redirect tabs.
Hence, I moved all of the adding and removing functionality to the backend (background.js) and only send events from the front-end to notify the back-end when the user wants to add/remove a URL. The back-end then sends a response with the outcome and the UI is updated accordingly.