Created a wrapper around the chrome API calls
Created a browser module, which acts as a wrapper around the API calls. I've done this to make it easier later to port to browsers that don't use the chrome API (I only have to adjust the code in the wrapper) and to shorten the code and make it more readable.